+++ napari_tiff/_tests/test_tiff_reader.py
@@ -1,47 +1,18 @@
+import dask.array as da
 import numpy as np
-import os
 import pytest
-import tifffile
-import zipfile
 
 from napari_tiff import napari_get_reader
 from napari_tiff.napari_tiff_reader import (imagecodecs_reader,
-                                            tifffile_reader,
-                                            zip_reader)
-
-
-def example_data_filepath(tmp_path, original_data):
-    example_data_filepath = str(tmp_path / "myfile.tif")
-    tifffile.imwrite(example_data_filepath, original_data)
-    return example_data_filepath
-
-
-def example_data_tiff(tmp_path, original_data):
-    filepath = str(tmp_path / "myfile.tif")
-    tifffile.imwrite(filepath, original_data)
-    return tifffile.TiffFile(filepath)
-
-
-def example_data_imagej(tmp_path, original_data):
-    filepath = str(tmp_path / "myfile.tif")
-    tifffile.imwrite(filepath, original_data, imagej=True)
-    return tifffile.TiffFile(filepath)
-
-
-def example_data_ometiff(tmp_path, original_data):
-    filepath = str(tmp_path / "myfile.ome.tif")
-    tifffile.imwrite(filepath, original_data, ome=True)
-    return tifffile.TiffFile(filepath)
-
-
-def example_data_zipped(tmp_path, original_data):
-    example_tiff_filepath = str(tmp_path / "myfile.tif")
-    tifffile.imwrite(example_tiff_filepath, original_data)
-    example_zipped_filepath = str(tmp_path / "myfile.zip")
-    with zipfile.ZipFile(example_zipped_filepath, 'w') as myzip:
-        myzip.write(example_tiff_filepath)
-    os.remove(example_tiff_filepath)  # not needed now the zip file is saved
-    return example_zipped_filepath
+                                            tifffile_reader, zip_reader)
+from napari_tiff._tests.test_data import (
+    example_data_filepath,
+    example_data_zipped_filepath,
+    example_data_tiff,
+    example_data_imagej,
+    example_data_ometiff,
+    example_data_multiresolution,
+)
 
 
 def test_get_reader_pass():
@@ -52,10 +23,10 @@ def test_get_reader_pass():
 
 @pytest.mark.parametrize("data_fixture, original_data", [
     (example_data_filepath, np.random.random((20, 20))),
-    (example_data_zipped, np.random.random((20, 20))),
+    (example_data_zipped_filepath, np.random.random((20, 20))),
     ])
 def test_reader(tmp_path, data_fixture, original_data):
-    """An example of how you might test your plugin."""
+    """Test tiff reader with example data filepaths."""
 
     my_test_file = data_fixture(tmp_path, original_data)
 
@@ -70,7 +41,7 @@ def test_reader(tmp_path, data_fixture, original_data):
     assert isinstance(layer_data_tuple, tuple) and len(layer_data_tuple) > 0
 
     # make sure it's the same as it started
-    if data_fixture == example_data_zipped:  # zipfile has unsqueezed dimension
+    if data_fixture == example_data_zipped_filepath:  # zipfile has unsqueezed dimension
         np.testing.assert_allclose(original_data, layer_data_tuple[0][0])
     else:
         np.testing.assert_allclose(original_data, layer_data_tuple[0])
@@ -81,7 +52,7 @@ def test_reader(tmp_path, data_fixture, original_data):
     (tifffile_reader, example_data_imagej, np.random.randint(0, 255, size=(20, 20)).astype(np.uint8)),
     (tifffile_reader, example_data_tiff, np.random.randint(0, 255, size=(20, 20)).astype(np.uint8)),
     (tifffile_reader, example_data_ometiff, np.random.randint(0, 255, size=(20, 20, 3)).astype(np.uint8)),
-    (zip_reader, example_data_zipped, np.random.random((20, 20))),
+    (zip_reader, example_data_zipped_filepath, np.random.random((20, 20))),
     ])
 def test_all_readers(reader, data_fixture, original_data, tmp_path):
     """Test each individual reader."""
@@ -96,7 +67,20 @@ def test_all_readers(reader, data_fixture, original_data, tmp_path):
     assert isinstance(layer_data_tuple, tuple) and len(layer_data_tuple) > 0
 
     # make sure it's the same as it started
-    if data_fixture == example_data_zipped:  # zipfile has unsqueezed dimension
+    if data_fixture == example_data_zipped_filepath:  # zipfile has unsqueezed dimension
         np.testing.assert_allclose(original_data, layer_data_tuple[0][0])
     else:
         np.testing.assert_allclose(original_data, layer_data_tuple[0])
+
+
+def test_multiresolution_image(example_data_multiresolution):
+    """Test opening a multi-resolution image."""
+    assert example_data_multiresolution.series[0].is_pyramidal
+    layer_data_list = tifffile_reader(example_data_multiresolution)
+    layer_data_tuple = layer_data_list[0]
+    layer_data = layer_data_tuple[0]
+    assert len(layer_data) == 3
+    assert layer_data[0].shape == (16, 512, 512, 3)
+    assert layer_data[1].shape == (16, 256, 256, 3)
+    assert layer_data[2].shape == (16, 128, 128, 3)
+    assert all([isinstance(level, da.Array) for level in layer_data])
